Bug fix: the delay was not initialised properly, hence there was random noise added to the output.
git-svn-id: svn+ssh://jop91.astron.nl/auto/home/jive_cvs/SOFTC/svn/trunk/sfxc@517 5736e3ea-d426-0410-8aaf-93e336ac3372

+++ include/data_reader.h
@@ -29,7 +29,7 @@ public:
       \return the number of bytes read into the buffer.
       \pre buff is allocated. nBytes >= 0
   **/
-  size_t get_bytes(size_t nBytes, char *buff);
+  int get_bytes(size_t nBytes, char *buff);
 
   /** Returns true if all data is read from the input reader.
   **/ 
@@ -62,7 +62,7 @@ public:
 private:
   /** Function that actually writes the data to the output device.
   **/ 
-  virtual size_t do_get_bytes(size_t nBytes, char *buff) = 0;
+  virtual int do_get_bytes(size_t nBytes, char *buff) = 0;
 
   uint64_t _data_counter;
   int data_slice;
